Splitted models module into sub-modules
