+++ src/models/site.rs
@@ -1,0 +1,8 @@
+
+#[derive(Queryable)]
+pub struct Site {
+    pub id: i32,
+    pub name: String,
+    pub url: String,
+    pub last_check: i32,
+}
